requirements.txt
requests
click
fastapi
pydantic-settings
uvicorn